Session-5-Responsible_Safe_AI/requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-core>=0.2.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
openpyxl>=3.1.2
pandas>=2.0.0
ddgs>=0.1.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
colorama>=0.4.6
tabulate>=0.9.0
streamlit>=1.35.0